+++ packages/tokens/src/constants/bsc.ts
@@ -3109,14 +3109,6 @@ export const bscTokens = {
     'Grape coin',
     'https://www.joingrapes.com/',
   ),
-  via: new ERC20Token(
-    ChainId.BSC,
-    '0x21ac3bB914f90A2Bb1a16088E673a9fdDa641434',
-    18,
-    'VIA',
-    'Octavia',
-    'https://octavia.one/',
-  ),
   ego: new ERC20Token(
     ChainId.BSC,
     '0x44a21B3577924DCD2e9C81A3347D204C36a55466',