+++ packages/price-api-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pancakeswap/price-api-sdk
 
+## 1.2.7
+
+### Patch Changes
+
+- @pancakeswap/smart-router@6.0.6
+
+## 1.2.6
+
+### Patch Changes
+
+- c1f1288: Update tyeps
+
 ## 1.2.5
 
 ### Patch Changes