+++ packages/widgets-internal/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pancakeswap/widgets-internal
 
+## 0.1.46
+
+### Patch Changes
+
+- Updated dependencies [a29edf6]
+  - @pancakeswap/v3-sdk@3.8.2
+  - @pancakeswap/farms@1.0.60
+
+## 0.1.45
+
+### Patch Changes
+
+- @pancakeswap/farms@1.0.59
+
 ## 0.1.44
 
 ### Patch Changes